+++ cirq-ionq/cirq_ionq/__init__.py
@@ -18,7 +18,7 @@ from cirq_ionq.calibration import Calibration
 
 from cirq_ionq.ionq_devices import IonQAPIDevice, decompose_to_device
 
-from cirq_ionq.ionq_gateset import IonQTargetGateset
+from cirq_ionq.ionq_gateset import IonQTargetGateset, decompose_all_to_all_connect_ccz_gate
 
 from cirq_ionq.ionq_exceptions import (
     IonQException,
